requirements.txt
Django==1.5.1
Markdown==2.3.1
Pillow==2.0.0
South==0.8.1
Unipath==1.0
argparse==1.2.1
boto==2.9.4
chardet==2.1.1
distribute==0.6.34
dj-database-url==0.2.1
django-admin-bootstrapped==0.4.1
django-appconf==0.6
django-boto==0.1.6b
django-ckeditor==4.0.2
django-csvimport==1.0
django-easy-maps==0.8.1
django-imagekit==3.0.0
django-markdown==0.2.1
django-storages==1.1.8
geopy==0.95.1
gunicorn==17.5
httplib2==0.8
pilkit==0.3.0
psycopg2==2.5
six==1.3.0
twilio==3.5.1
unittest2==0.5.1
wsgiref==0.1.2
